[layouts] Moves each layout spec to its own file
Details:
- Moves each layout spec to its own file in the layouts/collection
  folder
- Replaces custom default FA2 values with the ones from the
  graphology-layout-forceatlas2 repository (which in practice replaces
  default slowDown value from 10 to 1)
- Fixes autoSettings button, so that it no longer empties some of the
  setting inputs
